rename JS script
