+++ apps/site/assets/js/app.js
@@ -11,7 +11,7 @@ import googleMapsLoaded from "./google-maps-loaded";
 import submitOnEvents from "./submit-on-events";
 import selectModal from "./select-modal";
 import collapse from "./collapse";
-import mobileMenu from "../ts/app/mobile-menu";
+import setupGlobalNavigation from "../ts/app/global-navigation";
 import headerDropdowns from "./header-dropdowns";
 import modal from "./modal";
 import turbolinksMods from "./turbolinks-mods";
@@ -330,7 +330,7 @@ submitOnEvents(["blur", "change"]);
 googleAnalytics();
 googleMapsLoaded();
 selectModal();
-mobileMenu();
+setupGlobalNavigation();
 headerDropdowns();
 collapse();
 alertItemLoad();
