+++ apps/site/assets/ts/app/global-navigation.ts
