+++ apps/site/assets/ts/jest.config.js
@@ -9,7 +9,7 @@ module.exports = {
     "!**/*-entry.ts", // not necessary to test entry files
     "!**/*-loader.tsx", // not necessary to test loader files
     "!**/search.ts", // for now,
-    "!./app/mobile-menu.ts" // tested with Cypress
+    "!./app/global-navigation.ts" // tested with Cypress
   ],
   coverageReporters: ["html"],
   coverageThreshold: {
